USERS-PRODucts-Comments-carts CRUD done

+++ Models/Products.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Description { get; set; }
         public string Image { get; set; }
-        public decimal Pricing { get; set; }
-        public decimal ShippingCost { get; set; }
+        public int Pricing { get; set; }
+        public int ShippingCost { get; set; }
     }
 }
